// moonbowcl/src/app/model/dto/attr/attrEpayConf.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

export class AttrEpayConf {
    causale: string;
    codice_fiscale_ente: string;
    tipo_pagamento: string;
    importo: string;

    codice_fiscale_piva: string;
    cognome: string;
    nome: string;
    ragione_sociale: string;
    email: string;
}